+++ src/VideoDemos.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.0.31611.283
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VideoDemos", "VideoDemos\VideoDemos.csproj", "{881681C9-138A-4715-8927-1A9C01830BEC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VideoPlayback", "VideoPlayback\VideoPlayback.csproj", "{FDBD6319-E139-4190-BE3F-6AE16740406F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,6 +19,10 @@ Global
 		{881681C9-138A-4715-8927-1A9C01830BEC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{881681C9-138A-4715-8927-1A9C01830BEC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{881681C9-138A-4715-8927-1A9C01830BEC}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{FDBD6319-E139-4190-BE3F-6AE16740406F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FDBD6319-E139-4190-BE3F-6AE16740406F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FDBD6319-E139-4190-BE3F-6AE16740406F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FDBD6319-E139-4190-BE3F-6AE16740406F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
