+++ src/VideoDemos/Platforms/Android/MainActivity.cs
@@ -1,6 +1,5 @@
 ﻿using Android.App;
 using Android.Content.PM;
-using Android.OS;
 
 namespace VideoDemos;
