+++ src/VideoDemos/Platforms/MacCatalyst/Program.cs
@@ -1,5 +1,4 @@
-﻿using ObjCRuntime;
-using UIKit;
+﻿using UIKit;
 
 namespace VideoDemos;
