+++ src/VideoDemos/Platforms/Windows/App.xaml.cs
@@ -1,6 +1,4 @@
-﻿using Microsoft.UI.Xaml;
-
-// To learn more about WinUI, the WinUI project structure,
+﻿// To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
 
 namespace VideoDemos.WinUI;
