+++ src/VideoDemos/Views/CustomTransportPage.xaml.cs
@@ -1,5 +1,3 @@
-using VideoDemos.Controls;
-
 namespace VideoDemos.Views;
 
 public partial class CustomTransportPage : ContentPage
