+++ src/VideoDemos/Views/PlayLibraryVideoPage.xaml.cs
@@ -1,5 +1,3 @@
-using VideoDemos.Controls;
-
 namespace VideoDemos.Views;
 
 public partial class PlayLibraryVideoPage : ContentPage
