+++ src/VideoPlayback/Controls/IVideoController.cs
@@ -1,4 +1,4 @@
-﻿namespace VideoDemos.Controls
+﻿namespace VideoPlayback.Controls
 {
     public interface IVideoController
     {
