+++ src/VideoPlayback/Controls/VideoPositionEventArgs.cs
@@ -1,4 +1,4 @@
-﻿namespace VideoDemos.Controls
+﻿namespace VideoPlayback.Controls
 {
     public class VideoPositionEventArgs : EventArgs
     {
