+++ src/VideoPlayback/Controls/VideoSource.cs
@@ -1,8 +1,8 @@
 ﻿using System.ComponentModel;
 
-namespace VideoDemos.Controls
+namespace VideoPlayback.Controls
 {
-    
+
     [TypeConverter(typeof(VideoSourceConverter))]
     public abstract class VideoSource : Element
     {
