+++ src/VideoPlayback/Controls/VideoStatus.cs
@@ -1,4 +1,4 @@
-﻿namespace VideoDemos.Controls
+﻿namespace VideoPlayback.Controls
 {
     public enum VideoStatus
     {
