+++ src/VideoPlayback/Handlers/VideoHandler.MaciOS.cs
@@ -1,9 +1,9 @@
 ﻿#nullable enable
 using Microsoft.Maui.Handlers;
-using VideoDemos.Controls;
-using VideoDemos.Platforms.Windows;
+using VideoPlayback.Controls;
+using VideoPlayback.Platforms.MacCatalyst;
 
-namespace VideoDemos.Handlers
+namespace VideoDemos.Platforms.MaciOS
 {
     public partial class VideoHandler : ViewHandler<Video, MauiVideoPlayer>
     {
@@ -24,12 +24,12 @@ namespace VideoDemos.Handlers
 
         public static void MapAreTransportControlsEnabled(VideoHandler handler, Video video)
         {
-            handler.PlatformView?.UpdateTransportControlsEnabled();
+            handler?.PlatformView.UpdateTransportControlsEnabled();
         }
 
         public static void MapSource(VideoHandler handler, Video video)
         {
-            handler.PlatformView?.UpdateSource();
+            handler?.PlatformView.UpdateSource();
         }
 
         public static void MapIsLooping(VideoHandler handler, Video video)
@@ -39,7 +39,7 @@ namespace VideoDemos.Handlers
 
         public static void MapPosition(VideoHandler handler, Video video)
         {
-            handler.PlatformView?.UpdatePosition();
+            handler?.PlatformView.UpdatePosition();
         }
 
         public static void MapUpdateStatus(VideoHandler handler, Video video, object? args)
