+++ src/VideoPlayback/Handlers/VideoHandler.Windows.cs
@@ -1,13 +1,13 @@
 ﻿#nullable enable
 using Microsoft.Maui.Handlers;
-using VideoDemos.Controls;
-using VideoDemos.Platforms.Android;
+using VideoDemos.Platforms.Windows;
+using VideoPlayback.Controls;
 
 namespace VideoDemos.Handlers
 {
     public partial class VideoHandler : ViewHandler<Video, MauiVideoPlayer>
     {
-        protected override MauiVideoPlayer CreatePlatformView() => new MauiVideoPlayer(Context, VirtualView);
+        protected override MauiVideoPlayer CreatePlatformView() => new MauiVideoPlayer(VirtualView);
 
         protected override void ConnectHandler(MauiVideoPlayer platformView)
         {
