+++ src/VideoPlayback/Handlers/VideoHandler.cs
@@ -1,0 +1,37 @@
+﻿#if IOS || MACCATALYST
+using PlatformView = VideoPlayback.Platforms.iOS.MauiVideoPlayer;
+#elif ANDROID
+using PlatformView = VideoDemos.Platforms.Android.MauiVideoPlayer;
+#elif WINDOWS
+using PlatformView = VideoDemos.Platforms.Windows.MauiVideoPlayer;
+#elif (NETSTANDARD || !PLATFORM) || (NET6_0_OR_GREATER && !IOS && !ANDROID)
+using PlatformView = System.Object;
+#endif
+using Microsoft.Maui.Handlers;
+using VideoPlayback.Controls;
+
+namespace VideoDemos.Handlers
+{
+    public partial class VideoHandler
+    {
+        public static IPropertyMapper<Video, VideoHandler> PropertyMapper = new PropertyMapper<Video, VideoHandler>(ViewHandler.ViewMapper)
+        {
+            [nameof(Video.AreTransportControlsEnabled)] = MapAreTransportControlsEnabled,
+            [nameof(Video.Source)] = MapSource,
+            [nameof(Video.IsLooping)] = MapIsLooping,
+            [nameof(Video.Position)] = MapPosition
+        };
+
+        public static CommandMapper<Video, VideoHandler> CommandMapper = new(ViewCommandMapper)
+        {
+            [nameof(Video.UpdateStatus)] = MapUpdateStatus,
+            [nameof(Video.PlayRequested)] = MapPlayRequested,
+            [nameof(Video.PauseRequested)] = MapPauseRequested,
+            [nameof(Video.StopRequested)] = MapStopRequested
+        };
+
+        public VideoHandler() : base(PropertyMapper, CommandMapper)
+        {
+        }
+    }
+}
