+++ src/VideoPlayback/Platforms/Android/VideoProvider.cs
